refactor: Separate helpers in different classes
