+++ src/main/kotlin/io/ionic/libs/osfilesystemlib/controller/OSFLSTDirectoriesHelper.kt
@@ -1,0 +1,28 @@
+package io.ionic.libs.osfilesystemlib.controller
+
+import io.ionic.libs.osfilesystemlib.controller.internal.createDirOrFile
+import io.ionic.libs.osfilesystemlib.controller.internal.deleteDirOrFile
+import io.ionic.libs.osfilesystemlib.model.OSFLSTCreateOptions
+import io.ionic.libs.osfilesystemlib.model.OSFLSTDeleteOptions
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+
+class OSFLSTDirectoriesHelper {
+    /**
+     * Create a directory
+     *
+     * @param options options to create the directory
+     * @return success if the directory was created successfully, error otherwise
+     */
+    suspend fun createDirectory(options: OSFLSTCreateOptions): Result<Unit> =
+        withContext(Dispatchers.IO) { createDirOrFile(options, isDirectory = true) }
+
+    /**
+     * Delete or directory
+     *
+     * @param options options to delete the directory
+     * @return success if the directory was deleted successfully, error otherwise
+     */
+    suspend fun deleteDirectory(options: OSFLSTDeleteOptions): Result<Unit> =
+        withContext(Dispatchers.IO) { deleteDirOrFile(options) }
+}
